Add files via upload
Hello guys, here is a project am currently working on. Its a program that is aim at helping the educational system especially at the level of the universities, help all Professors and Students in the organization of classes and collection of marks after any sort of evaluation. Since it happen most often in Cameroonian Universities in particular. Any help will be the most welcome.